extract utility functions into module
